crawler_rdb/requirements.txt
requests
python-dotenv
psycopg2-binary
pandas
numpy
openpyxl
openai